avid/h264: Properly implement in-place mv rotation
Seemingly innocuous diff, but what's going on is that we're doing a
hardware motion vector merge & in-place transform by rotating the
addresses at which the hardware produces the current frame's motion
vectors (as an intermediate work tile) which it in turn *consumes* for
bidirectional references. The code before this happened to work by pure
chance alone until it, well, didn't.

  idx  L0  L1  slice_type
 ------------------------
 [ 0   0  -1   2] I  Output #0's mv into slot 0
 [ 1   1  -1   0] P  Output #1's mv into slot 1
 [ 2   2  -1   0] P  Output #2's mv into slot 2
 [ 3   3   1   1] B  Output #3's mv into slot 3 && ref #1's mv @ slot 1
 [ 4   4   1   1] B  Output #4's mv into slot 4 && ref #1's mv @ slot 1
 [ 5   5  -1   0] P  Output #5's mv into slot 5
 [ 6   6   2   1] B  Output #6's mv into slot 4 && ref #2's mv @ slot 2
 [ 7   7   2   1] B  Output #7's mv into slot 4 && ref #2's mv @ slot 2

Signed-off-by: Eileen Yoon <eyn@gmx.com>