codecs/h265: Add 32x32 scaling list DC coeff WAR
The spec is actually somewhat ambiguous on this. It's a [2][6] array
for 16x16 and 32x32 DC coefficients, but 32x32 only has 2/6 arrays
(index 0,3). 7.4.5, on !pred_mode_flag it's presumed 16, so FFmpeg
memsets the entire array to 16. But we don't iterate over dimensions
1,2,4,5 in the first place, and we need those zeroed. The spec violation
part is that we should set it to 16 on !pred_mode_flag for index (0,3),
but we apparently don't. But that's not the case for PPS, which makes
this really weird.

Signed-off-by: Eileen Yoon <eyn@gmx.com>